feat: better reorderable handle, cleanup unused files, refactor files to separate concerns
all that's left are the sidebars
