+++ lib/util/constants.dart
@@ -1,6 +1,16 @@
-import 'package:flutter/cupertino.dart';
+// ignore_for_file: public_member_api_docs
+
+import 'package:flutter/material.dart';
 
 const sidebarPaddingSize = 16.0;
 const sidebarPaddingStep = 4.0;
 const sidebarPaddingHorizontal =
     EdgeInsets.symmetric(horizontal: sidebarPaddingSize);
+
+const kSidebarWidth = 240.0;
+const kTopbarHeight = 52.0;
+const kGestureSize = 8.0;
+const kGextFieldWidth = 96.0;
+const kSnapRadius = 3.0;
+
+const kColorList = [Colors.red, Colors.green, Colors.blue, Colors.yellow];
