--- lib/widgets/canvas.dart
@@ -1,1 +1,0 @@
-part of '../main.dart';
