--- lib/widgets/selector_widget.dart
@@ -1,11 +1,0 @@
-part of '../main.dart';
-
-/// Marquee
-class SelectorWidget extends StatelessWidget {
-  const SelectorWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Listener(child: Stack());
-  }
-}
