+++ pubspec.yaml
@@ -28,7 +28,6 @@ dependencies:
   signals: ^2.1.10
   flutter_bloc: ^8.1.3
   uuid: ^4.3.1
-  dart_sort_queue: ^0.0.2+3
 
 dev_dependencies:
   flutter_test:
